Add tests for data_types tasks
