+++ Типы данных/data_types-multiply.js
@@ -1,0 +1,7 @@
+// В переменных a и b хранятся числа.
+// Написать программу, которая выводит
+// в консоль произведение и сумму этих чисел.
+
+export function multiply(a, b) {
+  console.log(a * b);
+}
